implement another approach

+++ day11.hpp
@@ -7,7 +7,7 @@ namespace Day11 {
 
 using namespace std;
 
-int solve_day11_pt1(const string &input);
+long solve_day11_pt1(const string &input);
 int solve_day11_pt2(const string &input);
 
 } // namespace Day11
